[refactor] move other document models in lib/models
