+++ lib/api/controllers/auth.js
@@ -28,7 +28,7 @@ const _ = require('lodash');
 const { NativeController } = require('./base');
 const User = require('../../models/security/user');
 const formatProcessing = require('../../core/auth/formatProcessing');
-const ApiKey = require('../../core/storage/models/apiKey');
+const ApiKey = require('../../models/storage/apiKey');
 const { IncomingMessage } = require('http');
 const { errors: {KuzzleError} } = require('kuzzle-common-objects');
