+++ lib/api/controllers/security.js
@@ -27,7 +27,7 @@ const
   Bluebird = require('bluebird'),
   { NativeController } = require('./base'),
   formatProcessing = require('../../core/auth/formatProcessing'),
-  ApiKey = require('../../core/storage/models/apiKey'),
+  ApiKey = require('../../models/storage/apiKey'),
   errorsManager = require('../../util/errors'),
   {
     Request,
