+++ lib/core/realtime/notifier.js
@@ -23,7 +23,11 @@
 
 const _ = require('lodash');
 const Bluebird = require('bluebird');
-const Notification = require('./document');
+const {
+  DocumentNotification,
+  ServerNotification,
+  UserNotification,
+} = require('../../models/notification');
 
 /**
  * @class NotifierController
@@ -114,7 +118,7 @@ class NotifierController {
         return this._dispatch(
           'notify:server',
           channels,
-          new Notification.Server(type, message),
+          new ServerNotification(type, message),
           connectionId);
       }
     }
@@ -408,7 +412,7 @@ class NotifierController {
     }
 
     if (channels.length > 0) {
-      const notification = new Notification.Document(
+      const notification = new DocumentNotification(
         request,
         scope,
         action,
@@ -446,7 +450,7 @@ class NotifierController {
     }
 
     if (channels.length > 0) {
-      const notification = new Notification.User(request, scope, content);
+      const notification = new UserNotification(request, scope, content);
 
       return this._dispatch('notify:user', channels, notification);
     }
