+++ lib/core/security/userRepository.js
@@ -26,7 +26,7 @@ const {v4: uuid} = require('uuid');
 const User = require('../../models/security/user');
 const Repository = require('../shared/repository');
 const errorsManager = require('../../util/errors');
-const ApiKey = require('../storage/models/apiKey');
+const ApiKey = require('../../models/storage/apiKey');
 const { Request } = require('kuzzle-common-objects');
 
 /**
