+++ lib/core/storage/bootstrap/internalIndexBootstrap.js
@@ -21,13 +21,12 @@
 
 'use strict';
 
-const
-  debug = require('../../../util/debug')('kuzzle:bootstrap:internalIndex'),
-  Bluebird = require('bluebird'),
-  SafeBootstrap = require('./safeBootstrap'),
-  ApiKey = require('../models/apiKey'),
-  crypto = require('crypto'),
-  errorsManager = require('../../../util/errors');
+const debug = require('../../../util/debug')('kuzzle:bootstrap:internalIndex');
+const Bluebird = require('bluebird');
+const SafeBootstrap = require('./safeBootstrap');
+const ApiKey = require('../../../models/storage/apiKey');
+const crypto = require('crypto');
+const errorsManager = require('../../../util/errors');
 
 class InternalIndexBootstrap extends SafeBootstrap {
   constructor (kuzzle, indexStorage) {
