+++ lib/core/storage/storageEngine.js
@@ -24,7 +24,7 @@
 const Elasticsearch = require('../../services/storage/elasticsearch');
 const errorsManager = require('../../util/errors').wrap('services', 'storage');
 const ClientAdapter = require('./clientAdapter');
-const BaseModel = require('./models/baseModel');
+const BaseModel = require('../../models/storage/baseModel');
 const Bluebird = require('bluebird');
 
 class IndexCache {
