+++ lib/models/notification/index.js
@@ -22,7 +22,7 @@
 'use strict';
 
 module.exports = {
-  Document: require('./DocumentNotification'),
-  Server: require('./ServerNotification'),
-  User: require('./UserNotification')
+  DocumentNotification: require('./document'),
+  ServerNotification: require('./server'),
+  UserNotification: require('./user')
 };
