+++ lib/models/storage/apiKey.js
@@ -21,8 +21,8 @@
 
 'use strict';
 
-const debug = require('../../../util/debug')('kuzzle:models:apiKey');
-const errorsManager = require('../../../util/errors');
+const debug = require('../../util/debug')('models:storage:apiKey');
+const errorsManager = require('../../util/errors');
 const BaseModel = require('./baseModel');
 const crypto = require('crypto');
