+++ test/core/security/userRepository.test.js
@@ -6,7 +6,7 @@ const sinon = require('sinon').createSandbox();
 const KuzzleMock = require('../../mocks/kuzzle.mock');
 const Repository = require('../../../lib/core/shared/repository');
 const User = require('../../../lib/models/security/user');
-const ApiKey = require('../../../lib/core/storage/models/apiKey');
+const ApiKey = require('../../../lib/models/storage/apiKey');
 const UserRepository = require('../../../lib/core/security/userRepository');
 const {
   errors: {
